+++ test/coredotfinance/database/apikey.txt
@@ -1,0 +1,1 @@
+administrator:coredottoday208
